Decrement the iter by one (1) to counteract “for” loop increment.
I update the iter inside the for loop to skip a whole subtree. The original algorithm assumes that if we modify the iter, the next loop starts without incrementing it. However my C implementation increments iter no matter what. This causes the algorithm to skip the subtree AND the next node. I decrement iter by one (1) to prevent skipping a node.